requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn

face_alignment
python_speech_features
numba
resampy
#pyaudio
soundfile==0.12.1
einops
configargparse

lpips==0.1.3
imageio-ffmpeg

transformers==4.46.2
edge_tts
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate

librosa
openai
websockets==12.0
gradio_client
azure-cognitiveservices-speech

# RVM Background Removal Dependencies
torchvision
